requirements.txt
pip==25.2
absl-py==2.3.0
aiosqlite==0.21.0
altair==5.5.0
altex==0.2.0
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.39.3
botocore==1.39.3
cachetools==5.5.2
captcha==0.7.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
decorator==5.2.1
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
entrypoints==0.4
extra-streamlit-components==0.1.80
fabric==3.2.2
Faker==37.4.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
favicon==0.7.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
gspread==6.2.1
h11==0.16.0
htbuilder==0.9.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib-metadata==6.11.0
invoke==2.2.0
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mediapipe==0.10.11
ml_dtypes==0.5.1
narwhals==1.44.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.3.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
orjson==3.10.18
package-name==0.1
packaging==23.2
pandas==2.3.0
paramiko==3.5.1
passlib==1.7.4
Pillow==9.5.0
platformdirs==4.3.8
plotly==6.2.0
prometheus_client==0.22.1
proto-plus==1.26.1
protobuf==3.20.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pymdown-extensions==10.16
Pympler==1.1
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==13.9.1
rich-toolkit==0.14.8
rignore==0.5.1
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.0
scipy==1.16.0
sentencepiece==0.2.0
sentry-sdk==2.32.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.16.0
snowflake-snowpark-python==1.33.0
sortedcontainers==2.4.0
sounddevice==0.5.2
soupsieve==2.7
SQLAlchemy==2.0.41
st-annotated-text==4.0.2
st-theme==1.2.3
starlette==0.47.1
streamlit==1.40.1
streamlit-authenticator==0.4.2
streamlit-avatar==0.1.3
streamlit-calendar==1.3.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-drawable-canvas==0.9.3
streamlit-embedcode==0.1.2
streamlit-extras==0.1.5
streamlit-image-coordinates==0.3.1
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
tenacity==8.5.0
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==4.3.1
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
validators==0.35.0
watchdog==5.0.3
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2
zipp==3.23.0
streamlit
mediapipe
opencv-python-headless
